feat(player): refactor MoviePlayer and introduce MovieHero and MovieCast components
