+++ frontend/src/app/(main)/movies/[id]/page.tsx
@@ -2,10 +2,10 @@ import { notFound } from 'next/navigation'
 import Link from 'next/link'
 import { ArrowLeft } from 'lucide-react'
 import { getAccessToken } from '@/lib/session'
-import { MoviePlayer } from '@/components/page/MoviePlayer'
+import { MovieHero } from '@/components/Player/MovieHero'
+import { MovieCast } from '@/components/Player/MovieCast'
+import { MoviePlayer } from '@/components/Player/MoviePlayer'
 import { CommentSection } from '@/components/page/CommentSection'
-import { FavoriteButton } from '@/components/page/FavoriteButton'
-import { WatchLaterButton } from '@/components/page/WatchLaterButton'
 import type { Metadata } from 'next'
 
 interface CastMember {
@@ -68,18 +68,18 @@ async function getMovie(id: string): Promise<MovieDetail | null> {
   }
 }
 
-export async function generateMetadata({ params }: { params: Promise<{ id: string }> }): Promise<Metadata> {
-  const { id } = await params
-  const movie = await getMovie(id)
-  if (!movie) return { title: 'Film introuvable — Hypertube' }
-  return { title: `${movie.title} (${movie.year}) — Hypertube` }
-}
-
 function formatRuntime(minutes: number): string {
   if (!minutes) return ''
   const h = Math.floor(minutes / 60)
   const m = minutes % 60
   return h > 0 ? `${h}h ${m.toString().padStart(2, '0')}m` : `${m}m`
+}
+
+export async function generateMetadata({ params }: { params: Promise<{ id: string }> }): Promise<Metadata> {
+  const { id } = await params
+  const movie = await getMovie(id)
+  if (!movie) return { title: 'Film introuvable — Hypertube' }
+  return { title: `${movie.title} (${movie.year}) — Hypertube` }
 }
 
 export default async function MovieDetailPage({ params }: { params: Promise<{ id: string }> }) {
@@ -97,85 +97,20 @@ export default async function MovieDetailPage({ params }: { params: Promise<{ id
         <ArrowLeft className="size-4" />
       </Link>
 
-      {movie.backdrop_url && (
-        <div className="relative w-full h-56 sm:h-72 md:h-96 overflow-hidden">
-          <img
-            src={movie.backdrop_url}
-            alt=""
-            className="w-full h-full object-cover"
-          />
-          <div className="absolute inset-0 bg-gradient-to-t from-background via-background/60 to-transparent" />
-        </div>
-      )}
-
       <div className="max-w-5xl mx-auto px-4 pb-16">
-        <div className={`flex gap-6 ${movie.backdrop_url ? '-mt-24 relative z-10' : 'mt-8'}`}>
-          {movie.poster_url && (
-            <div className="hidden sm:block shrink-0 w-36 md:w-48">
-              <img
-                src={movie.poster_url}
-                alt={movie.title}
-                className="w-full rounded-lg shadow-lg border border-border"
-              />
-            </div>
-          )}
+        <MovieHero
+          id={movie.id}
+          title={movie.title}
+          year={movie.year}
+          runtime={formatRuntime(movie.runtime)}
+          rating={movie.rating}
+          overview={movie.overview}
+          genres={movie.genres ?? []}
+          posterUrl={movie.poster_url}
+          backdropUrl={movie.backdrop_url}
+        />
 
-          <div className="flex flex-col gap-3 justify-end min-w-0">
-            <h1 className="text-2xl md:text-4xl font-bold leading-tight">{movie.title}</h1>
-
-            <div className="flex flex-wrap items-center gap-3 text-sm text-muted-foreground">
-              {movie.year && <span>{movie.year}</span>}
-              {movie.runtime > 0 && <span>{formatRuntime(movie.runtime)}</span>}
-              {movie.rating > 0 && (
-                <span className="text-destructive font-semibold">⭐ {movie.rating.toFixed(1)}</span>
-              )}
-            </div>
-
-            {(movie.genres?.length ?? 0) > 0 && (
-              <div className="flex flex-wrap gap-2">
-                {movie.genres.map(genre => (
-                  <span
-                    key={genre}
-                    className="text-xs px-2 py-0.5 rounded-full border border-border bg-muted text-muted-foreground"
-                  >
-                    {genre}
-                  </span>
-                ))}
-              </div>
-            )}
-
-            {movie.overview && (
-              <p className="text-sm text-muted-foreground leading-relaxed max-w-2xl">{movie.overview}</p>
-            )}
-
-            <div className="flex items-center gap-2 flex-wrap">
-              <FavoriteButton tmdbId={movie.id} />
-              <WatchLaterButton tmdbId={movie.id} />
-            </div>
-          </div>
-        </div>
-
-        {(movie.cast?.length ?? 0) > 0 && (
-          <section className="mt-10">
-            <h2 className="text-lg font-semibold mb-3">Casting</h2>
-            <div className="flex gap-3 overflow-x-auto pb-2">
-              {movie.cast.slice(0, 12).map(member => (
-                <div
-                  key={`${member.name}-${member.order}`}
-                  className="shrink-0 w-24 text-center"
-                >
-                  <div className="w-16 h-16 rounded-full bg-muted mx-auto flex items-center justify-center overflow-hidden">
-                    <span className="text-xl font-bold text-muted-foreground">
-                      {member.name.charAt(0)}
-                    </span>
-                  </div>
-                  <p className="text-xs font-medium mt-2 truncate">{member.name}</p>
-                  <p className="text-xs text-muted-foreground truncate">{member.character}</p>
-                </div>
-              ))}
-            </div>
-          </section>
-        )}
+        <MovieCast cast={movie.cast ?? []} />
 
         <section className="mt-10">
           <MoviePlayer torrents={movie.torrents ?? []} movieId={movie.id} />
